requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.2
fastapi==0.101.1
fonttools==4.42.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.2
multidict==6.0.4
nltk==3.8.1
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==2.2.1
pydantic_core==2.6.1
pymongo==4.4.1
pyparsing==3.0.9
pypdf==3.15.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
